[chore] :clown_face: move countries outside mocks (#129)
