+++ src/app/api/_data/_mock/_others.ts
@@ -1,36 +1,5 @@
 // ----------------------------------------------------------------------
 
-export const _socials = [
-  {
-    value: 'facebook',
-    name: 'FaceBook',
-    icon: 'eva:facebook-fill',
-    color: '#1877F2',
-    path: 'https://www.facebook.com/caitlyn.kerluke'
-  },
-  {
-    value: 'instagram',
-    name: 'Instagram',
-    icon: 'ant-design:instagram-filled',
-    color: '#E02D69',
-    path: 'https://www.instagram.com/caitlyn.kerluke'
-  },
-  {
-    value: 'linkedin',
-    name: 'Linkedin',
-    icon: 'eva:linkedin-fill',
-    color: '#007EBB',
-    path: 'https://www.linkedin.com/caitlyn.kerluke'
-  },
-  {
-    value: 'twitter',
-    name: 'Twitter',
-    icon: 'eva:twitter-fill',
-    color: '#00AAEC',
-    path: 'https://www.twitter.com/caitlyn.kerluke'
-  }
-]
-
 // ----------------------------------------------------------------------
 
 export const _homePlans = [...Array(3)].map((_, index) => ({
