+++ src/layouts/main/footer.tsx
@@ -13,8 +13,7 @@ import { usePathname } from 'src/routes/hooks'
 import { RouterLink } from 'src/routes/components'
 
 import { useTranslate } from 'src/locales'
-import { _socials } from 'src/app/api/_data/_mock'
-import { BOT_WAPP_URL, CONTACT_EMAIL } from 'src/config-global'
+import { _socials, BOT_WAPP_URL, CONTACT_EMAIL } from 'src/config-global'
 
 import Logo from 'src/components/logo'
 import Iconify from 'src/components/iconify'
